Removing duplicate code.
