+++ src/main/java/uk/gov/hmcts/reform/civil/model/dq/Respondent1DQ.java
@@ -6,17 +6,12 @@ import lombok.Data;
 import lombok.Setter;
 import uk.gov.hmcts.reform.civil.enums.YesOrNo;
 import uk.gov.hmcts.reform.civil.model.StatementOfTruth;
-import uk.gov.hmcts.reform.civil.model.UnavailableDate;
-import uk.gov.hmcts.reform.civil.model.UnavailableDateLRspec;
 import uk.gov.hmcts.reform.civil.model.account.AccountSimple;
 import uk.gov.hmcts.reform.civil.model.common.Element;
 import uk.gov.hmcts.reform.civil.model.documents.Document;
-import uk.gov.hmcts.reform.civil.utils.ElementUtils;
 
-import java.util.Collections;
 import java.util.List;
 import java.util.Optional;
-import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
 @Setter
@@ -92,46 +87,17 @@ public class Respondent1DQ implements DQ {
         if (respondent1DQHearing != null) {
             return getHearing(respondent1DQHearing);
         }
+        DQUtil util = new DQUtil();
+
         if (respondent1DQHearingFastClaim != null) {
-            return Hearing.builder()
-                .hearingLength(respondent1DQHearingFastClaim.getHearingLength())
-                .hearingLengthDays(respondent1DQHearingFastClaim.getHearingLengthDays())
-                .hearingLengthHours(respondent1DQHearingFastClaim.getHearingLengthHours())
-                .unavailableDatesRequired(respondent1DQHearingFastClaim.getUnavailableDatesRequired())
-                .unavailableDates(mapDates(respondent1DQHearingFastClaim.getUnavailableDatesLRspec()))
-                .build();
+            return util.buildFastTrackHearing(respondent1DQHearingFastClaim);
         }
         if (respondent1DQHearingSmallClaim != null) {
             SmallClaimHearing small = getSmallClaimHearing();
-            return Hearing.builder()
-                .unavailableDatesRequired(small.getUnavailableDatesRequired())
-                .unavailableDates(mapDates(small.getSmallClaimUnavailableDate()))
-                .build();
+            return util.buildSmallClaimHearing(small);
         }
 
         return null;
-    }
-
-    private List<Element<UnavailableDate>> mapDates(List<Element<UnavailableDateLRspec>> lrDates) {
-        if (lrDates == null) {
-            return Collections.emptyList();
-        } else {
-            return lrDates.stream().map(Element::getValue)
-                .map(this::mapDate)
-                .map(ElementUtils::element)
-                .collect(Collectors.toList());
-        }
-    }
-
-    private UnavailableDate mapDate(UnavailableDateLRspec lrSpec) {
-        UnavailableDate.UnavailableDateBuilder builder = UnavailableDate.builder()
-            .who(lrSpec.getWho());
-        if (lrSpec.getDate() != null) {
-            builder.date(lrSpec.getDate());
-        } else {
-            builder.fromDate(lrSpec.getFromDate()).toDate(lrSpec.getToDate());
-        }
-        return builder.build();
     }
 
     @Override
